+++ app/Http/Controllers/Admin/MemberController.php
@@ -6,7 +6,7 @@ use Illuminate\Http\Request;
 
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
-use App\Team;
+use App\Models\Team;
 
 class MemberController extends Controller
 {
